tank-combat: size chapter-2 Level-1 distance to a byte (prove + measure the fit)
The within-screen all-pairs distance was stored as uint16 sized to the
0..299 theoretical worst case. But the table is built from the grid —
build-time-known data, not adversarial input — so size it to what it can
actually take:

- It provably fits a byte: a shortest path never revisits a cell and no
  2x2 block lies wholly on an induced path (four cells close a loop), so
  the longest in-screen distance <= N_CELLS - (GRID_W/2)*(GRID_H/2) - 1 =
  229 < 255. A _Static_assert (grid_paths.h) guarantees it at compile time.
- l1_build_* now also returns the *measured* max (l1_maxdist /
  l1_screen_max), surfaced on the page, so an overflow would be detected,
  not silently truncated.

Result: l1dist 1.41 MB -> 706 KB, the dominant table; chapter 2 drops back
under chapter 1's 1 MiB (from 4 MiB), still no dynamic allocation.

Adds a byte-fit test (provable ceiling + measured max + a forced serpentine
screen). 40 checks pass; cross-screen routing and wall-edit rebuilds
unchanged. Refines the AGENTS.md "size to the real domain" rule with the
build-time prove/measure/detect corollary.

+++ tank-combat/chapter-2/src/edge_paths.c
@@ -71,7 +71,7 @@ void l2_build(World* w) {
     if (pj != 0xFF && 1u < D2(w, i, pj)) { D2(w, i, pj) = 1; NH(w, i, pj) = pj; }
   }
   for (uint32_t s = 0; s < N_SCREENS; s++) {          /* same-screen hops cost the Level-1 distance */
-    const uint16_t* L1s = w->l1dist + (uint32_t)s * TRI;
+    const uint8_t* L1s = w->l1dist + (uint32_t)s * TRI;
     uint32_t cnt = w->screen_ep_count[s];
     for (uint32_t a = 0; a < cnt; a++)
       for (uint32_t b = 0; b < cnt; b++) {
@@ -106,7 +106,7 @@ void l2_compute_pg(World* w, uint32_t p) {
   if (!w->phas[p]) return;
 
   uint32_t ds = w->pdest_screen[p], dc = w->pdest_cell[p];
-  const uint16_t* L1d = w->l1dist + ds * TRI;
+  const uint8_t* L1d = w->l1dist + ds * TRI;
   uint32_t cnt = w->screen_ep_count[ds];
   for (uint32_t y = 0; y < P; y++) {
     uint32_t best = INF32;
@@ -126,7 +126,7 @@ void l2_compute_pg(World* w, uint32_t p) {
 uint8_t route_next_dir(const World* w, uint32_t p, uint32_t screen, uint32_t cell) {
   if (!w->phas[p]) return DIR_NONE;
   uint32_t ds = w->pdest_screen[p], dc = w->pdest_cell[p];
-  const uint16_t* L1s = w->l1dist + screen * TRI;
+  const uint8_t* L1s = w->l1dist + screen * TRI;
 
   if (screen == ds) {
     if (cell == dc) return DIR_NONE;                     /* arrived */
